pyproject.toml
[project]
name = "patch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "honcho>=2.0.0",
    "huey>=2.5.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai[examples,logfire]>=0.2.9",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
